SVN_SILENT warnings--
svn path=/trunk/extragear/multimedia/amarok/; revision=598337

+++ src/statusbar/statusbar.cpp
@@ -263,7 +263,7 @@ StatusBar::drawTimeDisplay( int ms )  //SLOT
         s.prepend( '-' );
     }
 
-    while( s.length() < m_timeLength )
+    while( (int)s.length() < m_timeLength )
         s.prepend( ' ' );
     s += ' ';
